[REF] Board 초기화로직을 별도 BoardInitializer 클래스로 추출
